Pull request for merging visualisation into master

+++ project1/src/main/java/scheduler/ReducedStateScheduler.java
@@ -27,7 +27,7 @@ public class ReducedStateScheduler implements Scheduler {
 
     private int numTasks, numProcessors;
     private int bound = NO_SOLUTION; // The current best solution.
-    private int totalStates = 0, completeStates = 0, activeBranches = 0;
+    private long totalStates = 0, completeStates = 0, activeBranches = 0;
     private TaskGraph input;
 
     private int[] topologicalOrder;
@@ -207,18 +207,18 @@ public class ReducedStateScheduler implements Scheduler {
     }
 
     @Override
-    public int getTotalStatesVisited() {
+    public long getTotalStatesVisited() {
         return totalStates;
     }
 
     @Override
-    public int getCompleteStatesVisited() {
+    public long getCompleteStatesVisited() {
         return completeStates;
     }
 
     @Override
     public int getActiveBranches() {
-        return activeBranches;
+        return (int)activeBranches;
     }
 
     /**
@@ -348,7 +348,6 @@ public class ReducedStateScheduler implements Scheduler {
                     state.unassignedTasks++;
                     state.endTime = prevEndTime;
                 }
-
             }
 
             activeBranches--;
